+++ packages/flowr_arch/lib/src/fr_repo_impl.dart
@@ -104,14 +104,14 @@ abstract class FrRepo<T extends FrTable> extends IRepo<T, String> {
   ///   true: the record is only created if it does not exist.;
   ///   false: always create
   @override
-  Future<String> create(T value, {bool ifNotExists = true}) async {
+  Future<T> create(T value, {bool ifNotExists = true}) async {
     final r = table.record(value.id);
     await r.put(
       databaseClient,
       value.toJson()..remove('id'),
       ifNotExists: ifNotExists,
     );
-    return value.id;
+    return value;
   }
 
   @override