+++ packages/flowr_arch/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0 2025-8-27
+* feat 
+  - **break** IRepo::create will return DTO, not ID
+  - IRepo::deleteAll; FrRepo::deleteAll;
+* fix FrRepo::updateBy can't return DTO WyattHu(胡文涛) Yesterday 17:46
+
 ## 0.1.2 2025-8-26
 * adp Dart 3.7
 * feat IRepo::deleteAll; FrRepo::deleteAll;